ujimaru-markov-experiment/pyproject.toml
[tool.poetry]
name = "ujimaru-markov-experiment"
version = "0.1.0"
description = ""
authors = ["hppRC <hpp.ricecake@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyknp = "^0.4.1"
markovify = "^0.8.0"
mecab-python3 = "^0.996.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
